Add Prettier plugin for solidity files

+++ contracts/lib/CompoundOracleInterface.sol
@@ -3,9 +3,7 @@ pragma solidity ^0.5.0;
 contract CompoundOracleInterface {
     // returns asset:eth -- to get USDC:eth, have to do 10**24/result,
 
-
-    constructor() public {
-    }
+    constructor() public {}
 
     /**
   * @notice retrieves price of an asset
@@ -13,9 +11,8 @@ contract CompoundOracleInterface {
   * @param asset Asset for which to get the price
   * @return uint mantissa of asset price (scaled by 1e18) or zero if unset or contract paused
   */
-    function getPrice(address asset) public view returns (uint);
+    function getPrice(address asset) public view returns (uint256);
     // function getPrice(address asset) public view returns (uint) {
     //     return 527557000000000;
     // }
-
 }
