+++ contracts/lib/MockCompoundOracle.sol
@@ -1,15 +1,14 @@
 pragma solidity 0.5.10;
 
 contract MockCompoundOracle {
-    constructor() public {
-    }
+    constructor() public {}
 
-    uint256 price = 200 * (10 ** 18);
-    function getPrice(address asset) public view returns (uint) {
+    uint256 price = 200 * (10**18);
+    function getPrice(address asset) public view returns (uint256) {
         return price;
     }
 
     function updatePrice(uint256 newPrice) public {
-        price = newPrice * (10 ** 18);
+        price = newPrice * (10**18);
     }
 }
